Rename net/http.* to http_server.*
